+++ src/App.js
@@ -7,13 +7,13 @@ import Tricots from "./components/tricots/Tricots";
 import * as settingsActions from './actions/settingsActions';
 import store from './store';
 import HomeContainer from './containers/HomeContainer';
-import SettingsContainer from './containers/SettingsContainer';
-import RankingsContainer from './containers/RankingsContainer';
+import SettingsContainer from './containers/settings/SettingsContainer';
+import RankingsContainer from './containers/rankings/RankingsContainer';
 import GlobalHeaderContainer from './containers/GlobalHeaderContainer';
-import JudgmentsContainer from './containers/JudgmentsContainer';
-import Notifications from "./containers/NotificationsContainer";
+import JudgmentsContainer from './containers/judgments/JudgmentsContainer';
+import Notifications from "./containers/notifications/NotificationsContainer";
 import NoMatch from "./components/NoMatch";
-import TricotsContainer from './containers/TricotsContainer'
+import TricotsContainer from './containers/tricots/TricotsContainer'
 
 class App extends Component {
     state = {activeItem: 'home'}
