+++ src/components/RiderDetail/RiderTrikot.js
